backend/pyproject.toml
[project]
name = "myfinance-backend"
version = "0.1.0"
description = "MyFinance 통합 자산 관리 백엔드"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncmy>=0.2.9",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "google-search-results>=2.4.2",
    "yfinance>=0.2.0",
    "firecrawl-py>=1.0.0",
    "tavily-python>=0.5.0",
    "cryptography>=42.0.0",
    "litellm>=1.0.0",
    "sse-starlette>=1.6.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
